% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfg_colorscheme.R
\name{adfg_paletter}
\alias{adfg_paletter}
\title{Interpolate the selected palette}
\usage{
adfg_paletter(palette = "glacier", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Palette name in "adfg_palettes"}

\item{reverse}{TRUE/FALSE of palette order}

\item{...}{Additional arguments}
}
\value{
Characters, hex values of colors
}
\description{
Interpolate the selected palette
}
\examples{
adfg_paletter("glacier")(7)
}
