% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfg_colorscheme.R
\name{colorlister}
\alias{colorlister}
\title{Combine colors from a list}
\usage{
colorlister(...)
}
\arguments{
\item{...}{Color names}
}
\value{
named character
}
\description{
Given a list of color names & hex colors, returns hex colors as a list
}
\examples{
colorlister("auroragreen")
}
