% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfg_colorscheme.R
\name{display_palette}
\alias{display_palette}
\title{Create a display box output of the selected palette}
\usage{
display_palette(name, n, ...)
}
\arguments{
\item{name}{Palette name}

\item{n}{Number of colors to display}

\item{...}{Other arguments}
}
\value{
Graphic device
}
\description{
Create a display box output of the selected palette
}
\examples{
display_palette("glacier", 7)
}
