% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfg_colorscheme.R
\name{exactpal}
\alias{exactpal}
\title{Returns exact palette select, no interpolatio}
\usage{
exactpal(pal = "totem")
}
\arguments{
\item{pal}{Palette name}
}
\value{
Character
}
\description{
This function just returns the same colors as listed, in the order listed
To be used only when "useexact" is called
}
