% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfg_colorscheme.R
\name{scale_color_adfg}
\alias{scale_color_adfg}
\title{Color scale helper to add directly to ggplot}
\usage{
scale_color_adfg(
  palette = "logo",
  discrete = TRUE,
  reverse = FALSE,
  useexact = FALSE,
  ...
)
}
\arguments{
\item{palette}{Palette name.}

\item{discrete}{TRUE/FALSE of whether aesthetic is discrete}

\item{reverse}{TRUE/FALSE of palette order}

\item{useexact}{TRUE/FALSE of whether to use palettes exactly, no
interpolation}

\item{...}{Other arguments}
}
\value{
The output returned will be either an object of
\code{scale_color_manual()}, \code{discrete_scale()}, or
\code{scale_color_gradientn()}
}
\description{
\code{scale_color_adfg} adds a color scale directly to an existing ggplot.
Currently, continuous palette options are: "glacier", "sitkasunset",
"aurora", "alpenglow", "rockfish", "blues2", "tongass", "nigaq", "redblue",
"grays", "grays_bw". Discrete palette options are: "totem", "logo", "nome",
"berry", "bristolbay", "camai", "coho", and "denali". For discrete scales, to
use exact color palettes without interpolation, use argument \code{useexact =
TRUE}.
}
\examples{
\dontrun{
 ggplot(mpg, aes(x = hwy, y = cty, color = displ)) +
 geom_point() +
 scale_color_adfg(palette = "glacier", discrete = FALSE)
}
}
\seealso{
\code{\link{scale_fill_adfg}}
}
