% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfg_colorscheme.R
\name{scale_fill_adfg}
\alias{scale_fill_adfg}
\title{Fill scale helper to add directly to ggplot}
\usage{
scale_fill_adfg(
  palette = "logo",
  discrete = TRUE,
  reverse = FALSE,
  useexact = FALSE,
  ...
)
}
\arguments{
\item{palette}{Palette name}

\item{discrete}{TRUE/FALSE of whether aesthetic is discrete}

\item{reverse}{TRUE/FALSE of palette order}

\item{useexact}{TRUE/FALSE of whether to use palettes exactly, no interpolation}

\item{...}{Other arguments}
}
\value{

}
\description{
Fill scale helper to add directly to ggplot
}
\examples{
\dontrun{
 ggplot(mpg, aes(manufacturer, fill = manufacturer)) +
 geom_bar(color = "black") +
 scale_fill_adfg(palette = "glacier", discrete = TRUE)
}
}
