% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfg_colorscheme.R
\name{scale_fill_adfg}
\alias{scale_fill_adfg}
\title{Fill scale helper to add directly to ggplot}
\usage{
scale_fill_adfg(
  palette = "logo",
  discrete = TRUE,
  reverse = FALSE,
  useexact = FALSE,
  ...
)
}
\arguments{
\item{palette}{Palette name}

\item{discrete}{TRUE/FALSE of whether aesthetic is discrete}

\item{reverse}{TRUE/FALSE of palette order}

\item{useexact}{TRUE/FALSE of whether to use palettes exactly, no
interpolation}

\item{...}{Other arguments}
}
\value{
The output returned will be either an object of
\code{scale_fill_manual()}, \code{discrete_scale()}, or
\code{scale_fill_gradientn()}
}
\description{
\code{scale_fill_adfg} adds a fill scale directly to an existing ggplot.
Currently, continuous palette options are: "glacier", "sitkasunset",
"aurora", "rockfish", "bristolbay", "grays", "grays_bw". Discrete palette
options are: "totem", "logo", "tetrad", "berries", "bristolbay", "discrete7".
For discrete scales, to use exact color palette without interpolation, use
argument \code{useexact = TRUE}.
}
\examples{
\dontrun{
 ggplot(mpg, aes(manufacturer, fill = manufacturer)) +
 geom_bar(color = "black") +
 scale_fill_adfg(palette = "glacier", discrete = TRUE)
}
}
\seealso{
\code{\link{scale_color_adfg}}
}
